refactor : move components into large_componenents to prepare the small components and improve readability
