+++ client/src/components/large_components/GameControlsComponent.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect, JSX } from "react";
 import { useLocation } from "react-router-dom";
-import Dices from "./dices/HeroQuestDicesComponent";
+import Dices from "../dices/HeroQuestDicesComponent";
 import "./GameControlsComponent.css";
 import {
   Accordion,
@@ -9,20 +9,20 @@ import {
   Tooltip,
   Typography,
 } from "@mui/material";
-import { getMonsterIconPath } from "../shared/utils";
-import RedDices from "./dices/RedDicesComponent";
-import { monsterClassFr } from "../shared/languages/frenchEnums";
+import { getMonsterIconPath } from "../../shared/utils";
+import RedDices from "../dices/RedDicesComponent";
+import { monsterClassFr } from "../../shared/languages/frenchEnums";
 import MasterControls from "./MasterControlsComponent";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-import { getEquipmentName } from "../shared/equipments";
-import { TileType } from "../POO/enums/TileType";
-import { MonsterCategory } from "../POO/enums/Categories/MonsterCategory";
-import { GameAsJson } from "../POO/interfaces/ClassAsJson/Server/GameAsJson";
-import { Direction } from "../POO/enums/Direction";
-import { HeroAsJson } from "../POO/interfaces/ClassAsJson/Unit/HeroAsJson";
-import { MonsterAsJson } from "../POO/interfaces/ClassAsJson/Unit/MonsterAsJson";
-import { getHeroesByPlayerId, getPlayerIdToPlay } from "../shared/serverUtils";
-import { PlayerRole } from "../POO/enums/PlayerRole";
+import { getEquipmentName } from "../../shared/equipments";
+import { TileType } from "../../POO/enums/TileType";
+import { MonsterCategory } from "../../POO/enums/Categories/MonsterCategory";
+import { GameAsJson } from "../../POO/interfaces/ClassAsJson/Server/GameAsJson";
+import { Direction } from "../../POO/enums/Direction";
+import { HeroAsJson } from "../../POO/interfaces/ClassAsJson/Unit/HeroAsJson";
+import { MonsterAsJson } from "../../POO/interfaces/ClassAsJson/Unit/MonsterAsJson";
+import { getHeroesByPlayerId, getPlayerIdToPlay } from "../../shared/serverUtils";
+import { PlayerRole } from "../../POO/enums/PlayerRole";
 
 interface GameControlsProps {
   socket: any;
