+++ client/src/components/large_components/StatsComponent.tsx
@@ -4,20 +4,18 @@ import {
   getFightDiceFaceNumber,
   getIconClassPath,
   getUnitClassName,
-  isHero,
-} from "../shared/utils";
+} from "../../shared/utils";
 import { useEffect, useState } from "react";
 import { Socket } from "socket.io-client";
-import { getEquipmentName } from "../shared/equipments";
-import { Position } from "../POO/classes/Position/Position";
-import { MonsterAsJson } from "../POO/interfaces/ClassAsJson/Unit/MonsterAsJson";
-import { HeroAsJson } from "../POO/interfaces/ClassAsJson/Unit/HeroAsJson";
-import { StatsAsJson } from "../POO/interfaces/ClassAsJson/Unit/StatsAsJson";
+import { PositionAsJson } from "../../POO/interfaces/ClassAsJson/PositionAsJson";
+import { MonsterAsJson } from "../../POO/interfaces/ClassAsJson/Unit/MonsterAsJson";
+import { HeroAsJson } from "../../POO/interfaces/ClassAsJson/Unit/HeroAsJson";
+import { StatsAsJson } from "../../POO/interfaces/ClassAsJson/Unit/StatsAsJson";
 
 interface StatsComponentProps {
   socket: Socket;
   gameId: string;
-  position: Position;
+  position: PositionAsJson;
   unit: MonsterAsJson | HeroAsJson;
   setStatsVisible: (arg0: boolean) => void;
   isGameMaster: boolean;
